[bin] Fix binary location procedure to work with symlinks.
In #7860 `Sys.executable_name` was introduced to locate coq private
binaries; however, this breaks use cases with symlinks, such as Dune.

In order to fix this, we adopt a simpler strategy; we will always
adapt the name originally provided by the user, and only add a
directory if it was originally present.

This should work (hopefully) in all cases.

This fixes `coqide` not working on Dune builds.