[kernel] Add preliminary trace operators.
We'd need to add support for import for sure, but likely it is better
to hook at a higher level so we can get the module name instead of the
contents and replay from there.

We also have to add trace operators in lib.ml due to the
implementation of sections.