[kernel] Don't re-declare scheme side-effects that are already there.
This is an experimental PR as I am not sure I can follow the reasoning
in e1ba72037191b1d4be9de8a0a8fc1faa24eeb12c

Note that in safe_typing we avoid re-declaring effects twice.

This removes a huge number of redeclaration of schemes side-effects
that in fact are already generated such as `eq_ind`.

Anyways we should deprecate the declaration of Schemes on-the-fly, I
don't see the point honestly; just make sure your theory has the right
ones.

Well, going to a more eager declaration scheme could be costly in
terms of size, so OMMV.

TODO: only declare side-effects if the scheme is generated on-the-fly,
add a parameter to the declaration function.