[obligations] Pre-functionalize Program state
In our quest to unify all the declaration paths, an important step
is to account for the state pertaining to `Program` declarations.

Whereas regular proofs keep are kept in a stack-like structure;
obligations for constants defined by `Program` are stored in a global
map which is manipulated by almost regular open/close proof primitives.

State-threading in plugins is quite complicated, but we commit this PR
in preparation for the switch to a purely functional state in #11836