[record] Move back libobject manipulation to Record ops
This reverts some bits of ff07d0f499dcc8876b2b222bf861e9de89315a05
(still some improvements done there remain), in order to remove
internals from `recordops.mli`.

What to do here is indeed a tricky point, and we wouldn't like to have
calls to `Global.env ()` at this level, however, there still remain a
few points in the pretyper and more generally among libobject users.

It seems to me then that we will forced anyways to provide the right
environment from the libobject API [I think that's not hard to do], as
there are many other users of this patterns in the rest of the system.
These calls should be eliminated then easily.