[ltac debug] General refactoring towards a self-contained protocol
We refactor the LTAC debugger interface as to have a more standard
interface, in particular we improve error handling and parsing API
so it doesn't rely on tricky exception handling, and add an ADT for
debugger answers. We also handle some anomalies properly.

Critically, we remove the overloading of feedback to implement the
debugger prompt and we instead use a dedicated channel.

This commit should get the code into a mergeable state.

TODO:

- finish hooks / xml serialization
- finish coqide integration
- check changelog for XML protocol