[wip] Remove use of low-level entry objects in favor of higher-level API.
This is just an attempt to use the higher-level API in a few places,
unfortunately this is far from optimal hence the draft status so
hopefully we can see what would be the best setup.

It is still unclear if we want to add an API shortcut as to allow
declaration of terms that are evar-free, but still the API manages the
universe stuff.