[library] Allow require to load libraries from memory.
This is one more step in our quest to avoid saving / loading .vo
files to disk when not needed.

Most changes are straightforward except for the extension of the
command type to include an internalization function.

This is IMHO a good start to the goal of having the vernacular layer
access all file-system by a parameterized structure to interpretation,
so document managers can properly handle file-system access.

For example, a typical use case in jsCoq is to download a .vo file on
Require, but there are many more possibilities.