[build] [doc] Fix coq-doc package
This uses Dune 3.5 support for installing directories to make the opam
package workable.

This is for now a draft, but I have tried with these two scenarios:

- `dune build coq-doc.install # dev build`
- `dune build -p coq-doc`     # opam build`

The dune bump could be handled in a regular fashion (just bump to 3.5
for 8.19), or by having a separate dune-project for the docs.